+++ static/src/constants.js
@@ -49,6 +49,8 @@ export const TSNE_CONFIG = {
 };
 export const STEPS_TO_MOVE_TO_SPACE = 1;
 
+export const DEFAULT_TRUNCATED_STRING_LENGTH = 40;
+
 export const START_METRONOME_AT_MOUNT = false;
 export const DEFAULT_TEMPO = 120.0;
 export const LOOKAHEAD = 25; // How often we'll call the scheduler function (in milliseconds)
