+++ static/src/containers/Map/utils.js
@@ -1,0 +1,6 @@
+import { sidebarWidth } from 'json!../../stylesheets/variables.json';
+
+export const getMapCenter = () => ({
+  x: parseInt(sidebarWidth, 10) + ((window.innerWidth - parseInt(sidebarWidth, 10)) / 2),
+  y: (window.innerHeight / 2),
+});
