+++ static/src/utils/colorsUtils.js
